+++ braid/src/interface/oracle/utils.rs
@@ -247,6 +247,8 @@ pub fn gen_sobol_samples(
 
 // Weight Calculation
 // ------------------
+// Returns `given_weights`. `given_weights[s][v]` is the mixture weights for
+// view v in state s.
 #[allow(clippy::ptr_arg)]
 #[inline]
 pub fn given_weights(
@@ -345,10 +347,14 @@ fn single_view_weights(
 
     match given {
         Given::Conditions(ref conditions) => {
-            for &(id, ref datum) in conditions {
-                let in_target_view = state.asgn.asgn[id] == target_view_ix;
+            for &(col_ix, ref datum) in conditions {
+                let in_target_view = state.asgn.asgn[col_ix] == target_view_ix;
                 if in_target_view {
-                    view.ftrs[&id].accum_weights(datum, &mut weights, false);
+                    view.ftrs[&col_ix].accum_weights(
+                        datum,
+                        &mut weights,
+                        false,
+                    );
                 }
             }
             let z = logsumexp(&weights);
@@ -755,6 +761,8 @@ macro_rules! dep_ind_col_mixtures {
     ($states: ident, $col_a: ident, $col_b: ident, $fx: ident) => {{
         let mut mms_dep = Vec::new();
         let mut mms_ind = Vec::new();
+
+        // Proportion of mass in states in which col_a and col_b are dependent
         let mut weight = 0.0;
         $states.iter().for_each(|state| {
             let mm = match state.feature_as_mixture($col_a) {
@@ -791,7 +799,8 @@ pub fn categorical_gaussian_entropy_dual(
     let (_, cm_dep, cm_ind) =
         dep_ind_col_mixtures!(states, col_cat, col_gauss, Categorical);
 
-    // The number of components in a categorical model should never exceed u8
+    // The number of values the categorical column can take on, e.g. if the
+    // categories are "Yes", "No", and "Maybe", cat_k should be 3.
     let cat_k = match states[0].feature(col_cat) {
         ColModel::Categorical(cm) => u8::try_from(cm.prior.k())
             .expect("Categorical k exceeded u8 max value"),
@@ -821,6 +830,7 @@ pub fn categorical_gaussian_entropy_dual(
     // pass empty containers where they're not expected.
     let has_dep_states = gm_dep.k() > 0;
     let has_ind_states = gm_ind.k() > 0;
+    let has_ind_and_dep_states = has_ind_states && has_dep_states;
 
     // order of the polynomial for gauss-legendre quadrature
     let quad_level = 16;
@@ -851,9 +861,10 @@ pub fn categorical_gaussian_entropy_dual(
     // NOTE: this will take a really long time when k is large
     -(0..cat_k)
         .map(|k| {
-            // NOTE: I've chose to use the logp instead of vanilla 'p'. It
+            // NOTE: I've chosen to use the logp instead of vanilla 'p'. It
             // doesn't really change the runtime.
             let ind_cat_f = if has_ind_states {
+                // TODO: can cache this
                 cm_ind.ln_f(&k)
             } else {
                 // assert_eq!(dep_weight, 1.0);
@@ -869,7 +880,7 @@ pub fn categorical_gaussian_entropy_dual(
 
             let quad_fn = |y: f64| {
                 // We have to compute things differently for states in which the
-                // two columns are dependent and independent. The dependednt
+                // two columns are dependent and independent. The dependent
                 // computation is a bit more complicated.
                 let dep_cpnt = if has_dep_states {
                     let mut m = dep_cache.borrow_mut();
@@ -887,10 +898,20 @@ pub fn categorical_gaussian_entropy_dual(
                         .zip(ln_fys)
                         .map(|(w, g)| w + *g)
                         .collect();
-                    logsumexp(&cpnts)
+
+                    let ln_f = logsumexp(&cpnts);
+
+                    // If we do not have independent components, we do not have
+                    // to compute the weighted sum of the two scenarios, so we
+                    // can return early
+                    if !has_ind_and_dep_states {
+                        return ln_f * ln_f.exp();
+                    } else {
+                        ln_f
+                    }
                 } else {
                     assert_eq!(dep_weight, 0.0);
-                    1.0
+                    1.0 // ln(0) = 1
                 };
 
                 // We can basically cache the entire independent computation, so
@@ -900,12 +921,22 @@ pub fn categorical_gaussian_entropy_dual(
                     let mut m = ind_cache.borrow_mut();
                     let ln_fy =
                         m.entry(F64::new(y)).or_insert_with(|| gm_ind.ln_f(&y));
-                    ind_cat_f + *ln_fy
+                    let ln_f = ind_cat_f + *ln_fy;
+
+                    // If we do not have dependent components, we do not have to
+                    // compute the weighted sum of the two scenarios, so we can
+                    // return early
+                    if !has_ind_and_dep_states {
+                        return ln_f * ln_f.exp();
+                    } else {
+                        ln_f
+                    }
                 } else {
                     assert_eq!(dep_weight, 1.0);
-                    0.0
+                    1.0 // ln(0) = 1
                 };
 
+                // Add the weighted dependent and independent mixture components
                 let ln_f = logsumexp(&[
                     dep_weight.ln() + dep_cpnt,
                     (1.0 - dep_weight).ln() + ind_cpnt,
@@ -916,6 +947,7 @@ pub fn categorical_gaussian_entropy_dual(
 
             let last_ix = points.len() - 1;
 
+            // right tail of integral
             let q_a = gauss_legendre_quadrature_cached(
                 quad_fn,
                 (lower, points[0]),
@@ -923,6 +955,7 @@ pub fn categorical_gaussian_entropy_dual(
                 &gl_cache.1,
             );
 
+            // left tail of integral
             let q_b = gauss_legendre_quadrature_cached(
                 quad_fn,
                 (points[last_ix], upper),
@@ -930,6 +963,7 @@ pub fn categorical_gaussian_entropy_dual(
                 &gl_cache.1,
             );
 
+            // interior integral points
             let q_m = if points.len() == 1 {
                 0.0
             } else {