+++ braid/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.39.5
+- Minor optimization in joint entropy computation between one categorical and
+    one continuous column
+
 ## 0.39.4
 - Fixed bug that caused engines saved with `id_offset` (including using `braid
     run -o <OFFSET>` to save duplicate states with the wrong indices.